pyproject.toml
[project]
name = "mcp-template"
version = "0.1.0"
description = "Cursor-friendly template for MCP servers with contract-first mcp.yaml workflow."
readme = "README.MD"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "mcp[cli]>=1.4.0",
    "uvicorn>=0.30.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = []

[dependency-groups]
dev = []
